% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cds}
\alias{cds}
\title{Cds rstan script.}
\format{An object of class \code{character} of length 1.}
\usage{
cds
}
\description{
script to perform a cds analysis with rstan.
}
\keyword{datasets}
