% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_varName.R
\name{change_effort_varName}
\alias{change_effort_varName}
\title{Standardisation of column names}
\usage{
change_effort_varName(effort_base)
}
\arguments{
\item{effort_base}{Data.frame containing column names to standardize.}
}
\value{
This function return the same data.frame as in input with the standardized column name.
}
\description{
This function allows to attribute standard names to column names of effort data.frame
for the next steps of the analysis.
}
\note{
When "strate_sec" is not found at the end of the function, it creates a strate_sec
column automatically corresponding to half of transect column and survey colum separated by "_".
      Example, for a row, if \code{transect = O2/306} and \code{survey = FR_N_C} we will get
      \code{strate_sec = FR_N_c_02}.
}
\examples{


}
