% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_varName.R
\name{change_obs_varName}
\alias{change_obs_varName}
\title{Standardisation of observation data column names.}
\usage{
change_obs_varName(obs_base)
}
\arguments{
\item{obs_base}{Data.frame containing column names to standardize.}
}
\value{
This function return the same data.frame as in input with the standardized column name.
}
\description{
Attribute standard names to column names of observation data.frame
for the next steps of the analysis.
}
