% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_geffaeR_project.R
\name{create_cds_project}
\alias{create_cds_project}
\title{Create cds project}
\usage{
create_cds_project(
  path = ".",
  rstudio = rstudioapi::isAvailable(),
  open = TRUE
)
}
\arguments{
\item{...}{}
}
\value{
Create a project environment to perform a cds analysis.
}
\description{
Create cds project
}
