% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DSM_pack_MOLMOL}
\alias{DSM_pack_MOLMOL}
\title{Data list to create a DSM example}
\format{
An object of class \code{list} of length 6.
}
\usage{
DSM_pack_MOLMOL
}
\description{
DSM list that gather a list of data needed to perform A DSM analysis.
}
\keyword{datasets}
