% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{effort_example}
\alias{effort_example}
\title{Effort example data.}
\format{
An object of class \code{data.frame} with 278 rows and 15 columns.
}
\usage{
effort_example
}
\description{
Effort data simulated in atlantic ocean in the bay of Biscay.
}
\keyword{datasets}
