% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Europe}
\alias{Europe}
\title{Europe land contour.}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 59 rows and 12 columns.
}
\usage{
Europe
}
\description{
Europe land contour for maps.
}
\keyword{datasets}
