% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all_dsm.R
\name{fit_all_dsm}
\alias{fit_all_dsm}
\title{Fit all possible model with the covariates given.}
\usage{
fit_all_dsm(
  distFit = NULL,
  segdata_obs,
  obsdata,
  response = "ind",
  predictors,
  likelihood = "negbin",
  esw = NULL,
  max_cor = 0.5,
  nb_max_pred = 3,
  complexity = 4,
  smooth_xy = TRUE,
  k = 5,
  splines_by = NULL,
  weighted = FALSE,
  random = NULL,
  soap = list(xt = NULL, knots = NULL),
  use_loo = FALSE
)
}
\arguments{
\item{distFit}{\code{distfit} object. One of \code{plot_detection} output or directly an output of \code{ds}}

\item{segdata_obs}{segdata data.frame with observation added with \code{ajout_obs}.}

\item{obsdata}{Observation data.frame prepared with \code{prepare_data_obs}.}

\item{response}{Response variable to choose between "ind" number of indiviudals or "obs" number of observation.}

\item{predictors}{vector containing all covariates as character string.}

\item{likelihood}{Likelihood must be one of "negbin", "poisson" or "tweedie". Default is "negbin".}

\item{esw}{Value of \emph{effective-strip width}.}

\item{max_cor}{Maximum correlation threshold between two covariates tolerated.}

\item{nb_max_pred}{Maximum number of covariates in a model.}

\item{complexity}{k argument in smoothers.}

\item{smooth_xy}{Controls the intercept to include or not a bivariate smooth on x and y :
for prediction inside the prospected polygon, should be set to TRUE to obtain stable estimates for
prediction outside, MUST be set to FALSE to keep extrapolation under control.}

\item{k}{Number of models to return for inference.}

\item{splines_by}{Interaction with splines given by one variable of segdata.}

\item{weighted}{Either to compute weights or no.}

\item{random}{variable of \code{segdata_obs} to include random effect.}

\item{soap}{list to pass in order to use a soap-film smooth: must be prepared outside this function.}

\item{use_loo}{Model selection with leave-one-out cross validation if TRUE, if FALSE it uses AIC to select models.}
}
\value{
This function return a list containing:
        \enumerate{
          \item all_fits_binded : \code{data.frame} Containing all fitting infos on models.
          \item best_models : \code{list} k best models with standardized value.
          \item best_models4plotting : \code{list} k best models with raw values (for plotting).
        }
}
\description{
Fit all possible model with the covariates given.
}
\note{
By default, use cubic B-splines with shrinkage (bs = 'cs').
}
