% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_variomodel.R
\name{fit_variomodel}
\alias{fit_variomodel}
\title{Fitting a variogram model.}
\usage{
fit_variomodel(variogram, form = c("matern", "exponential"))
}
\arguments{
\item{variogram}{Object of class variogram \code{\link[geoR]{variog}}.}

\item{form}{Shape of the adjusted function. Could be either "matern" or "exponential".}
}
\value{
vario_model : object of class "variomodel" and "variofit, see details on the \code{\link[geoR]{variofit}} page
}
\description{
Fitting a variogram model from a "variogram" object.
}
