% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcmc_summary.R
\name{get_mcmc_summary}
\alias{get_mcmc_summary}
\title{summarize mcmc output of stan}
\usage{
get_mcmc_summary(x, alpha = 0.2, dig = 4, median = FALSE)
}
\arguments{
\item{x}{Output of stan function}

\item{alpha}{Risk taken for \code{\link[coda]{HPDinterval} function.}}

\item{dig}{Digits to round output of the funtion.}

\item{median}{Logical. If TRUE, indexes will be around median (mad). If FALSE, indexes will be around
mean (sd).}
}
\description{
This function transform output of stan model in data.frame which is easier to use.
}
\examples{

}
