% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lbrt93_proj}
\alias{lbrt93_proj}
\title{Projection in Lambert 93.}
\format{
An object of class \code{character} of length 1.
}
\usage{
lbrt93_proj
}
\description{
It stores the Lambert 93 projection in CRS format to ease georeferenced transformation.
}
\keyword{datasets}
