% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEA}
\alias{NEA}
\title{Northern East Atlantic land contour.}
\format{An object of class \code{SpatialPolygons} of length 68.}
\usage{
NEA
}
\description{
Northern East Atlantic land contour for maps.
}
\keyword{datasets}
