% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEA_isobath100}
\alias{NEA_isobath100}
\title{Northern East Atlantic isobath100.}
\format{An object of class \code{SpatialLinesDataFrame} with 9 rows and 1 columns.}
\usage{
NEA_isobath100
}
\description{
Northern East Atlantic isobath for depth of 100m.
}
\keyword{datasets}
