% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEA_isobath200}
\alias{NEA_isobath200}
\title{Northern East Atlantic isobath200.}
\format{An object of class \code{SpatialLinesDataFrame} with 27 rows and 2 columns.}
\usage{
NEA_isobath200
}
\description{
Northern East Atlantic isobath for depth of 200m.
}
\keyword{datasets}
