% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEA_simplified_FR}
\alias{NEA_simplified_FR}
\title{Northern East Atlantic land contour simplified centered on France}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 0 columns.
}
\usage{
NEA_simplified_FR
}
\description{
Northern East Atlantic land contour for maps with a less important resolution for time economy...
}
\keyword{datasets}
