% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{observation_example}
\alias{observation_example}
\title{Observation example data.}
\format{An object of class \code{data.frame} with 130 rows and 18 columns.}
\usage{
observation_example
}
\description{
Effort data simulated in atlantic ocean in the bay of Biscay.
}
\keyword{datasets}
