% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_detection.R
\name{plot_detection}
\alias{plot_detection}
\title{distance smapling analysis}
\usage{
plot_detection(distdata, bin, key, upper = NULL)
}
\arguments{
\item{distdata}{distadata data.frame previously built with  \code{\link{prepare_data_obs}}.}

\item{bin}{Min and max value of the distance detection and the step between
theese two.}

\item{key}{Choice bewtween two key function :
\enumerate{
  \item "halfnorm" : Half-Normal.
  \item "hazard" : hazard-rate.
}}

\item{upper}{Maximum value of the distance range of detection estimation.}
}
\value{
This function return list with :
        \enumerate{
          \item esw : effective strip width. with confidence interval at 95 percent.
          \item esw_cv : coefficient of variation of esw in percent.
          \item graph : Barplot of detections depending of distance, with confindence
          estimated with mcmc chains.
        }
}
\description{
This function calculate the ESW (effective-strip witdth) of a taxonomic group,
It also use the \link[Distance]{ds} function to estimate all the abondance, density encounter-rate,...
parameter and summarize it in the results of the function.
Finally i gives the detection plot of the distance analysis.
}
\examples{


}
