% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_map_dsm.R
\name{pred_map_dsm}
\alias{pred_map_dsm}
\title{Prediction map of fitted dsm model.}
\usage{
pred_map_dsm(predata, grid, var, facet_param = NULL, title = NULL)
}
\arguments{
\item{predata}{data.frame of prediction.}

\item{grid}{gridata in \code{sf} format or \code{crs} value. It have to be usable by \code{\link[sf]{st_crs}}.}

\item{var}{Column name as character used as value to be plotted.}

\item{facet_param}{Column name as character used as facetting parameter.}

\item{title}{character string to put as title on the map.}
}
\value{
\code{ggplot} object.
}
\description{
Map of density surface modelling fitted with pred_splines.
}
