% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_splines.R
\name{pred_splines}
\alias{pred_splines}
\title{Prediction diagnostic.}
\usage{
pred_splines(
  segdata,
  dsm_model,
  remove_intercept = FALSE,
  random = NULL,
  splines_by = NULL,
  alpha = 0.8
)
}
\arguments{
\item{segdata}{segdata data.frame built with \code{prepare_data_effort}}

\item{dsm_model}{The dsm model you want to use, built with \code{fit_all_dsm}.}

\item{remove_intercept}{Remove intercept of model (\code{beta[, 1] <- 0}).}

\item{random}{Variable to include random effect.}

\item{splines_by}{Interaction with splines given by one variable of segdata.}

\item{alpha}{coverage level for confidence interval.}
}
\value{
\enumerate{
          \item df_splines : \code{data.frame}.
          \item g_splines : \code{ggplot}.
          \item spatial : \code{data.frame}.
        }
}
\description{
Visualize prediction fiiting and get
}
