% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_abund.r
\name{predict_abund}
\alias{predict_abund}
\title{Predict density from dsm models estimated with fit_all_dsm.}
\usage{
predict_abund(listdsm, predata, n_sim = 500, abund_by = NULL, alpha = 0.8)
}
\arguments{
\item{listdsm}{Whole object \code{fit_all_dsm}.}

\item{predata}{One of object built from \code{prep_predata} function corresponding to the data from which to predict.}

\item{n_sim}{number of draws to approximate posterior distribution with MV normal distribution.}

\item{abund_by}{grouping factor for abundance estimation.}

\item{alpha}{coverage level for confidence interval.}
}
\value{

}
\description{
Predict density from dsm models estimated with fit_all_dsm.
}
