% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dsm.R
\name{predict_all}
\alias{predict_all}
\title{Predict density from dsm models estimated with fit_all_dsm.}
\usage{
predict_all(listdsm, predata)
}
\arguments{
\item{listdsm}{Whole object \code{fit_all_dsm}.}

\item{predata}{One of object built from \code{prep_predata} function corresponding to the data from which to predict.}
}
\value{

}
\description{
Predict density from dsm models estimated with fit_all_dsm.
}
