% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_kriging.R
\name{predict_kriging}
\alias{predict_kriging}
\title{Get spatial density / Abundance prediction.}
\usage{
predict_kriging(
  segdata_obs,
  esw,
  vario_model,
  predict_coord_xy = NULL,
  fast_inversion = TRUE,
  saturate = TRUE,
  intraTransect = TRUE
)
}
\arguments{
\item{segdata_obs}{data.frame. Output of \code{\link[geffaeR]{ajout_obs}}.}

\item{esw}{Effective (half) Strip-Width determined by \code{\link[geffaeR]{plot_detection}}.}

\item{vario_model}{output of \code{\link[geffaeR]{fit_variomodel}}.}

\item{predict_coord_xy}{data.frame with X and Y columns for longitude and latitude,
giving area where the kriging prediction is made.}

\item{saturate}{If TRUE, allows to exclude extreme prediction by keeping values under 0.999 quantile}

\item{intraTransect}{Boolean. Consider only distance between point on the same segment for the kriging distance ?}
}
\value{
data.frame of mean prediction and its se (standard error) with coordinates (X and Y).
}
\description{
Give spatial density from adjusted covariance model.
}
