% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_predata.R
\name{prep_predata}
\alias{prep_predata}
\title{\encoding{Création de predata et imputation de données manquantes}}
\usage{
prep_predata(
  segdata,
  gridfile_name,
  varenviro,
  do_log_enviro,
  varphysio,
  do_log_physio,
  imputation = "Amelia",
  shape_pred,
  layer_pred,
  saturate_predata = F,
  saturate_segdata = F,
  inbox_poly = T
)
}
\arguments{
\item{segdata}{"Segdata" data.frame, built with \code{\link{prepare_data_effort}}.}

\item{gridfile_name}{\encoding{Grille de la zone d'étude avec toutes les variables d'intérêt qui permettront la fabrication
de predata.}}

\item{varenviro}{\encoding{Vecteur contenant les variables environnementales dynamiques (i.e chlorophylles, SST).}}

\item{do_log_enviro}{\encoding{vecteur précisant parmi les variables environnementales (varenviro), lesquelles
sont à transformer en logarithme néperien + 1 (\code{\link[base]{log1p}}).}}

\item{varphysio}{\encoding{Vecteur contenant les variables environnementales non-dynamiques (i.e depth, dist200m).}}

\item{do_log_physio}{\encoding{vecteur précisant parmi les variables physiologiques (varphysio), lesquelles
sont à transformer en logarithme néperien + 1 (\code{\link[base]{log1p}}).}}

\item{imputation}{\encoding{Méthode permettant l'imputation de données manquantes pour segdata et predata.
Parmi les méthodes il y a :
\itemize{
  \item{"PCA" : }{\encoding{par méthode ACP. méthode issue du package \pkg{missMDA}}.}
  \item{"Amelia" : }{\encoding{par méthode Amelia. méthode issue du pakage \pkg{Amelia}}.}
}
Par défaut la méthode est "PCA".}}

\item{shape_pred}{\encoding{shapefile délimitant la zone où la prédiction est effectuée.}}

\item{layer_pred}{\encoding{Layer du shpaefile donnant les limite de la zone de prédiction.}}

\item{proj_grid}{\encoding{Systême de projection du fichier de gride.}}
}
\value{
Cette fonction renvoit une liste contenant :
        \enumerate{
          \item predata : Un data.frame contenant predata
          \item segdata : Un data.frame contenant segdata
          \item pca_pred : Une sortie de la fonction \code{\link[FactoMineR]{PCA} sur le data.frame segdata}
          \item pca_seg : Une sortie de la fonction \code{\link[FactoMineR]{PCA} sur le data.frame predata}
        }
}
\description{
\encoding{Cette fonction permet de créer predata, le data.frame contenant les données environnementales dans la
grille de la zone d'étude. Cette fonction permet également l'imputation de données manquantes pour les data.frames
segdata et predata}
}
\examples{

}
