% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_obs.R
\name{prepare_data_obs}
\alias{prepare_data_obs}
\title{Preparation of observation data}
\usage{
prepare_data_obs(
  sp,
  obs_base,
  legdata,
  segdata,
  shape,
  shape_layer,
  projection,
  truncation = NULL,
  remove_sp = NULL,
  unit_km = FALSE
)
}
\arguments{
\item{sp}{Name of taxon, group, family or species.}

\item{obs_base}{Observation database.}

\item{legdata}{"Legdata" data.frame, built with \code{\link{prepare_data_effort}}.}

\item{segdata}{"Segdata" data.frame, built with \code{\link{prepare_data_effort}}.}

\item{shape}{\encoding{Nom du shape file de la zone d'étude.}}

\item{shape_layer}{\encoding{Couche du shape file d'interêt.}}

\item{projection}{Projection of \code{shape} object.}

\item{truncation}{Default = \code{NULL}. disstance of truncation of the sampling (in km)}

\item{remove_sp}{Species to remove of the data set.}

\item{unit_km}{Default = \code{FALSE}. Is the unit of distance in the data set is in km ?}
}
\value{
This function return a list containing :
        \enumerate{
          \item distdata : Data.frame with distance information, which aim to be an input for
          \code{\link[Distance]{ds}}.
          \item obsdata : Data.frame containing information at observation scale.
          \item countdata_leg : A data.frame that merge leg scale effort informations
          with number of sightings and number of individuals (N and Y).
          \item countdata_seg : A data.frame that merge segment scale effort informations
          with number of sightings and number of individuals (N and Y).
        }
}
\description{
This function transform raw observation data into multiple sub data.frame for next analysis of
other functions of the package.
}
\examples{


}
