% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_soap.R
\name{prepare_soap}
\alias{prepare_soap}
\title{Prepare soap for dsm modelling}
\usage{
prepare_soap(
  data,
  contour = NULL,
  polygon_pred = NULL,
  ratio_simplify = 1e-04,
  N = 23
)
}
\arguments{
\item{data}{data.frame containing at least 3 columns. Among those columns it must have \code{longitude} and
\code{latitude}.}

\item{contour}{Land contour of the study area where the soap will be implemented (optionnal).
Must be a Spatial or sf object containing polygons defining land contour (default is NULL).}

\item{polygon_pred}{Polygon defining the area where the soap will be implemented.
Either a Spatial or sf object containing polygon.}

\item{ratio_simplify}{Correspond to the 'keep' argument of \code{\link[rmapshaper]{ms_simplify}} (default 1e-4).}

\item{N}{Number of knots per row and column (default 23).}
}
\value{
\enumerate{
          \item sf_cropped_NEA : \code{data.frame} (given if \code{polygon_pred} is \code{NULL}).
          \item bnd : \code{list} of \code{list} delimiting soap area. Compatible with \code{\link[mgcv]{gam}}.
          \item knots : \code{data.frame} containing longitude and latitude of konts.
          \item df_cropped_data : \code{data.frame} corresponding to \code{data} without the points outside of \code{bnd}.
          \item gg_cropped_data : \code{gg} object that synthetised information on the soap process which has been made.
          \item sf_ocean : \code{sf} object containging the contour of the study area where knots are placed.
        }
}
\description{
Create necessary output for dsm modelling using soap smoother.
}
\references{
This function uses \code{autocruncher}, a function created by Simon Wood and improved by David L. Miller
(\link[gtihub link]{https://github.com/dill/soap_checker/blob/master/autocrunch.R}).
}
