% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_map_krig.R
\name{quick_map_krig}
\alias{quick_map_krig}
\title{Visualize kriging prediction.}
\usage{
quick_map_krig(df_sf, varname, segdata_obs = NULL)
}
\arguments{
\item{df_sf}{Object of class "sf", output of \code{\link[geffaeR]{predict_kriging}}.}

\item{varname}{Choose which column of df to plot between "mean_pred" or "se_pred".}

\item{segdata_obs}{Optionnal. Allows to plot observations on the map,
default is NULL. Output of \code{\link[geffaeR]{ajout_obs}}.}
}
\description{
Map projection of \code{\link[geffaeR]{predict_kriging}} output.
}
