% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shape_example}
\alias{shape_example}
\title{Shape-file example.}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 1 columns.
}
\usage{
shape_example
}
\description{
Shape-file of the example in atlantic ocean in the bay of Biscay.
}
\keyword{datasets}
