% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_cds.R
\name{stan_cds}
\alias{stan_cds}
\title{\encoding{Modelisation Baésienne à partir d'un modèle stan.}}
\usage{
stan_cds(cds_data, segdata_obs, distdata)
}
\arguments{
\item{cds_data}{\encoding{Modèle écrit en "stan". Le modèle doit être renseigné sous forme de character string.}}

\item{segdata_obs}{\encoding{data.frame segdata contenant les infos par segment avec le nombre de d'individus et le
nombre de détections. Provient de la fonction \code{\link{ajout_obs}}}}

\item{distdata}{\encoding{data.frame distdata. Provient de la fonction \code{\link{prepare_data_obs}}}}
}
\value{

}
\description{
\encoding{Cette fonction permet de d'appliquer un modèle bayésien sur les données
d'observation (distadata et segdata_obs) à partir d'un modèle stan}
}
\examples{


}
