% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram_intraTransect.R
\name{variogram_intraTransect}
\alias{variogram_intraTransect}
\title{Build a variogram object}
\usage{
variogram_intraTransect(
  segdata_obs,
  obs,
  esw,
  intraTransect = T,
  breaks = NULL,
  plot = TRUE,
  pairs.min = 100,
  n_sim = 100,
  distmat = NULL,
  region = NULL,
  esp = "species"
)
}
\arguments{
\item{segdata_obs}{data.frame. Output of \code{\link[geffaeR]{ajout_obs}}.}

\item{obs}{character. Choose what kind obs observation is analyzed, "n_ind" : number of individuals
counted or "n_detection" number of detection.}

\item{esw}{Effective (half) Strip-Width determined by \code{\link[geffaeR]{plot_detection}}.}

\item{intraTransect}{Boolean. Consider only distance between point on the same segment for the kriging distance ?}

\item{breaks}{Breaks point of the variogram. Default is NULL which create default breaks : \code{c(-0.01, 0.001, 1, seq(2.5, 100, 2.5))}.}

\item{plot}{Boolean. Get diagnostic plot in output (default is TRUE).}

\item{pairs.min}{(default is 100)}

\item{n_sim}{(default is 100)}

\item{distmat}{Object of class matrix giving distance matrix}

\item{region}{Region of segdata_obs on which adjusting function. Default is NULL,
meaning it takes all the regions in segdata_obs}

\item{esp}{Species name for graph title.}
}
\value{
Gives a list of :
  \enumerate{
    \item variogram : object of class variogram
    \item g : coefficient of variation of esw in percent.
    \item distmat : Barplot of detections depending of distance, with confindence
          estimated with mcmc chains.
  }
}
\description{
Create an object of class variogram (output of \code{\link[geoR]{variog}}) and gives distance matrix
and a diagnostic graph if asked.
}
