% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_varName.R
\name{change_effort_varName}
\alias{change_effort_varName}
\title{Standardisation of column names}
\usage{
change_effort_varName(effort_base)
}
\arguments{
\item{effort_base}{Data.frame containing column names to standardize.}
}
\value{
This function return the same data.frame as in input with the standardized column name.
}
\description{
This function allows to attribute standard names to column names of effort data.frame
for the next steps of the analysis.
}
\note{
When "strate_sec" is not found at the end of the function, it creates a strate_sec
      column automatically corresponding to the merge of "subRegion" and "strate". There are merged by "_".
      Example, for a row, if \code{strate = N1} and \code{subRegion = ATL} we will get
      \code{strateSec = ATL_N1}.
}
\examples{


}
