% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_detection.R
\name{plot_detection}
\alias{plot_detection}
\title{Distance sampling analysis.}
\usage{
plot_detection(distdata, bin, key, upper = NULL, is_seabird)
}
\arguments{
\item{distdata}{data.frame previously built with  \code{\link{prepare_data_obs}}.}

\item{bin}{Min and max value of the distance detection and the step between
these two.}

\item{key}{Choice bewtween two key function :
\enumerate{
  \item "halfnorm" : Half-Normal.
  \item "hazard" : hazard-rate.
}}

\item{upper}{Maximum value of the distance range of detection estimation.}
}
\value{
This function return list with :
        \enumerate{
          \item esw : effective strip width. with confidence interval at 95 percent.
          \item esw_cv : coefficient of variation of esw in percent.
          \item graph : Barplot of detections depending of distance, with confindence
          estimated with mcmc chains.
          \item distfit : distance object from \link[Distance]{ds} function.
        }
}
\description{
Adjust a detection function on distance observation for one taxonomic group.
It gives the ESW (effective-strip witdth), its standard deviation,
a distance object from \link[Distance]{ds} function,
and it gives the detection plot of the detection function adjusted.
}
