% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_predata.R
\name{prep_predata}
\alias{prep_predata}
\title{Creation of predata and data impuation of missing data.}
\usage{
prep_predata(
  segdata,
  gridfile_name,
  varenviro,
  do_log_enviro,
  varphysio,
  do_log_physio,
  imputation = "Amelia",
  shape,
  saturate_predata = F,
  saturate_segdata = F,
  inbox_poly = T,
  col2keep = NULL
)
}
\arguments{
\item{segdata}{"Segdata" data.frame, built with \code{\link{prepare_data_effort}}.}

\item{gridfile_name}{Grid of the study area with all the environmental variables which
allows to build predata.}

\item{varenviro}{Vector containing all dynamic environmental variables (e.g. chlorophyl, SST).}

\item{do_log_enviro}{Vector precising which variables among varenviro that are necessary
to transform to neperian logarithm + 1 (\code{\link[base]{log1p}}).}

\item{varphysio}{Vector containing non-dynamic environmental variables (i.e depth, dist200m).}

\item{do_log_physio}{Vector precising which variables among varphysio that are necessary
to transform to neperian logarithm + 1 (\code{\link[base]{log1p}}).}

\item{imputation}{Data imputation method to missing values of segdata and predata. 2 methods allowed :
\itemize{
  \item "PCA" : PCA method from \pkg{missMDA}.
  \item "Amelia" : Amelia method from \pkg{Amelia}.
}
Default method is PCA.}

\item{shape}{Shapefile of the study area. It can be either a \code{SpatialPolygonsDataFrame} class object,
in this case it is not necessary to give shape_layer argument. Or it can be the name of the shape
object with its extension ".shp" (ex : "data/studyAreaShapefile.shp").}

\item{saturate_predata}{Boolean. If \code{TRUE}, saturate function is applied on all varenviro
an varphysio columns of segdata. Saturate function excludes extreme valuesand keep values between
quantiles 95 percent and 5 percent.}

\item{saturate_segdata}{Boolean. If \code{TRUE}, saturate function is applied on all varenviro
an varphysio columns of predata. Saturate function excludes extreme valuesand keep values between
quantiles 95 percent and 5 percent.}

\item{inbox_poly}{When \code{TRUE}, keep only part of the grid that match with the study area.}

\item{col2keep}{\code{character string} corresponding to the columns wanted to appear in output in predata.}
}
\value{
This function return a list containing :
        \enumerate{
          \item predata : data.frame of predata.
          \item segdata : data.frame of segdata.
          \item pca_seg : Output of \code{\link[FactoMineR]{PCA}} function on predata.
          \item pca_pred : Output of \code{\link[FactoMineR]{PCA}} function on segdata.
          \item seg_mipat : output of VIM::aggr on predata
          \item pred_mipat : output of VIM::aggr on segdata
        }
}
\description{
Creation of predata, a data.frame containing environmental parameters in a grid of the study area.
Also it allows to impute missing data for predata and segdata.
}
\examples{

}
