% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_effort.R
\name{prepare_data_effort}
\alias{prepare_data_effort}
\title{Preparation of effort data}
\usage{
prepare_data_effort(
  effort_base,
  covariable = NULL,
  block_area,
  shape,
  shape_layer,
  New_projection,
  optimal = FALSE
)
}
\arguments{
\item{effort_base}{Data.frame with effort data.}

\item{covariable}{Vector of covariable names to keep in output of the function.}

\item{block_area}{Data.frame with 2 colnames :
\enumerate{
  \item Block.
  \item Area.
}}

\item{shape}{Shapefile of the study area.}

\item{shape_layer}{Layer of the shapefile.}

\item{New_projection}{New projection of longitude and latitude of columns
(POINT_X et POINT_Y) in Proj4String format,
see : \code{\link[sp]{CRS}} for more infos.}

\item{optimal}{Argument which allows to keep data sampled in optimal conditions.
Defaults settings are all data are kept. In case of optimal = T, indexes
\code{c("GG", "GM", "MG", "EG", "GE", "EE", "ME", "EM", "MM")} are kept.}
}
\value{
This function return :
        \enumerate{
          \item Legdata : Data.frame with infos at leg scale.
          \item Segdata : Data.frame with infos at segment scale.
        }
}
\description{
This function transform raw effort data into multiple sub data.frame for next analysis of
other functions of the package.
}
\examples{


}
