% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_effort.R
\name{prepare_data_effort}
\alias{prepare_data_effort}
\title{Preparation of effort data}
\usage{
prepare_data_effort(
  effort_base,
  covariable = NULL,
  block_area,
  shape,
  shape_layer,
  New_projection,
  optimal = FALSE
)
}
\arguments{
\item{effort_base}{\encoding{data.frame contenant les données d'effort.}}

\item{covariable}{\encoding{Vecteur contenant le nom des covariables à conserver en sortie de la fonction.}}

\item{block_area}{\encoding{data.frame contenant une colonne :
  \enumerate{
    \item Block.
    \item Area.
  }
}}

\item{shape}{\encoding{Nom du shape file de la zone d'étude.}}

\item{shape_layer}{\encoding{Couche du shape file d'interêt.}}

\item{New_projection}{\encoding{Nouvelle projection des longitude et latitude des
colonnes (POINT_X et POINT_Y) format Proj4String,
voir : \code{\link[sp]{CRS}} pour plus d'infos.}}

\item{optimal}{\encoding{Paramètre permettant de selectionner (ou non) les données prelevées dans les
conditions otpimales. Par défaut toutes les données sont selectionnées. Dans le cas où seules les
données en conditions optimales sont selectionnees, ce sont les indices
\code{c("GG", "GM", "MG", "EG", "GE", "EE", "ME", "EM", "MM")} qui sont conservés.}}
}
\value{
Cette fonction renvoie une liste  contenant :
        \enumerate{
          \item Legdata : Un data.frame contenant les infos sur chaque leg.
          \item Segdata : Un data.frame contenant les infos sur chaque segment.
        }
}
\description{
This function transform raw effort data into multiple sub data.frame for next analysis of
other functions of the package.
}
\examples{


}
