% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_cds.R
\name{stan_cds}
\alias{stan_cds}
\title{CDS Bayesian modelling from stan model.}
\usage{
stan_cds(cds_data, segdata_obs, distdata)
}
\arguments{
\item{cds_data}{`character string` corresponding of the stan model.}

\item{segdata_obs}{Segdata `data.frame` containing information at segment scale with the number
on individuals and number of detection added with \code{\link{ajout_obs}} function.}

\item{distdata}{`data.frame` distdata. Coming from \code{\link{prepare_data_obs}.}}
}
\value{

}
\description{
Apply a Bayesian model on observation data (distdata and segdata_obs) with a stan model.
}
\examples{


}
